EditButtonとDeleteButtonを削除しDialySales本体に実装
